+++ Exchange/ViewModel/CurrencyScrollViewModel.swift
@@ -11,9 +11,9 @@ class CurrencyScrollViewModel {
     var currentIndex: Variable<Int> = Variable<Int>(0)
     var currentItem: Variable<CurrencyType> = Variable<CurrencyType>(.eur)
     var sufficientFundsToExchange: Variable<Bool> = Variable<Bool>(true)
-    var storage: [CurrencyType: Variable<Float>] = [.eur: Variable<Float>(0),
-                                                    .gbp: Variable<Float>(0),
-                                                    .usd: Variable<Float>(0)]
+    var storage: [CurrencyType: Variable<Double>] = [.eur: Variable<Double>(0),
+                                                    .gbp: Variable<Double>(0),
+                                                    .usd: Variable<Double>(0)]
 
     init() {
         exchangeRateService.currenciesStorage.asObservable()
